requirements.txt
litellm
pydantic==2.7.1
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.6
